ALEGRIA_OS/backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
google-genai
prisma
groq
openai
httpx
duckduckgo-search>=6.0
beautifulsoup4
lxml
pydantic-settings
numpy
opencv-python
playwright
feedparser
